Add masks-counters
